Force int for param team_id

+++ team-edit.php
@@ -13,7 +13,7 @@ level_or_alert(3, 'Modification d’une équipe');
 $logged_id   = (int)$_SESSION['logged_id'];
 $logged_user = strtolower($_SESSION['logged_user']);
 
-$team_id = param_post_or_get('team_id', 0);
+$team_id = (int)param_post_or_get('team_id', 0);
 $mode = 'Modifier';
 if ($team_id == 0) // new
 	$mode = 'Ajouter';
